typecast error

+++ watdfs_client.cpp
@@ -339,7 +339,7 @@ int watds_read_write_full(
     int rpc_ret, fxn_ret;
     fxn_ret = 0;
     while (temp_left_to_read > 0) {
-        size_t to_read = std::min(MAX_ARRAY_LEN, temp_left_to_read);
+        size_t to_read = std::min((size_t)MAX_ARRAY_LEN, temp_left_to_read);
         rpc_ret = watdfs_read_write_single(userdata, path, temp_buf, to_read, temp_offset, fi, is_read);
         
         // error handling
